ref(dashboards): Update imports for renamed matchTimeSeriesToTableRowValue
Update import paths in the spec file and visualizationWidget to
reference the renamed module with corrected casing.
